requirements.txt
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocs-minify-plugin==0.6.4
mkdocs-redirects==1.2.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.37.1
python-dotenv==1.0.1
httpx==0.27.0